Create AddListActivity to add a new to-do list
If the user leaves the EditText blank, they receive an error stating
the name cannot be blank. If the user enters a name that is already in
use as a to-do list name, a Toast informs the user a new name must be
entered. The user also has the option to cancel without creating a list.